internal/testpty: move from os/signal/internal/pty
We're going to use this for another test, so make it more accessible.

Preparation for #37486.

Reviewed-on: https://go-review.googlesource.com/c/go/+/449502
Run-TryBot: Austin Clements <austin@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
